Support 32bit system

+++ line_profiler/timers.c
@@ -62,7 +62,7 @@ hpTimer(void)
         struct timespec ts;
         PY_LONG_LONG ret;
         clock_gettime(CLOCK_MONOTONIC, &ts);
-        ret = ts.tv_sec * 1000000000 + ts.tv_nsec;
+        ret = (int64_t)ts.tv_sec * 1000000000 + (int64_t)ts.tv_nsec;
         return ret;
 }
